Added `yarn setup` and `.perspectiverc` build env config script
