+++ package.json
@@ -15,20 +15,21 @@
         "@babel/polyfill": "^7.0.0",
         "@babel/preset-env": "^7.1.5",
         "arraybuffer-loader": "^1.0.2",
-        "base64-font-loader": "0.0.4",
         "babel-eslint": "^8.2.3",
         "babel-jest": "^24.5.0",
         "babel-plugin-lodash": "^3.3.4",
         "babel-plugin-transform-custom-element-classes": "^0.1.0",
         "babel-polyfill": "^6.26.0",
+        "base64-font-loader": "0.0.4",
         "chalk": "^2.4.2",
         "clean-css-cli": "4.2.1",
         "clean-css-loader": "^1.1.0",
-        "core-js": "^2.5.7",
-        "css-loader": "^0.28.7",
         "codemirror": "^5.30.0",
         "codemirror-javascript": "^0.2.0",
+        "core-js": "^2.5.7",
         "cpx": "^1.5.0",
+        "css-loader": "^0.28.7",
+        "dotenv": "^8.1.0",
         "eslint": "^4.19.1",
         "eslint-config-prettier": "^3.0.1",
         "eslint-plugin-prettier": "^2.6.2",
@@ -36,6 +37,7 @@
         "graceful-fs": "https://github.com/mekwall/node-graceful-fs.git#patch-1",
         "html-loader": "^0.5.1",
         "html-loader-jest": "^0.2.1",
+        "inquirer": "^7.0.0",
         "jest": "^24.5.0",
         "js-beautify": "^1.8.6",
         "jsdoc": "3.5.5",
@@ -52,6 +54,7 @@
         "prettier": "^1.14.2",
         "rimraf": "^2.6.2",
         "sinon": "^7.3.1",
+        "source-map-explorer": "^2.0.1",
         "style-loader": "^0.18.2",
         "ts-loader": "^6.0.4",
         "typescript": "~2.9.2",
@@ -72,6 +75,7 @@
         "install_python": "cd python/table && python3 setup.py install",
         "install_python:table": "cd python/perspective && python3 setup.py install",
         "bench": "node scripts/bench.js",
+        "setup": "node scripts/setup.js",
         "docs": "node scripts/docs.js",
         "test": "node scripts/test.js",
         "test_cpp": "node scripts/test_cpp.js",
